+++ shared/src/common/types/token/token-payload.type.ts
@@ -1,0 +1,7 @@
+type TokenPayload = {
+  data: string;
+  iat: number;
+  exp: number;
+};
+
+export { type TokenPayload };
